fix(exec): rename UnixExec to Exec and Exec to ExecBase to keep binary compatibility
