+++ src/main/java/com/aws/greengrass/util/platforms/Platform.java
@@ -9,7 +9,7 @@ import com.aws.greengrass.config.PlatformResolver;
 import com.aws.greengrass.logging.api.Logger;
 import com.aws.greengrass.logging.impl.LogManager;
 import com.aws.greengrass.util.CrashableFunction;
-import com.aws.greengrass.util.Exec;
+import com.aws.greengrass.util.ExecBase;
 import com.aws.greengrass.util.FileSystemPermission;
 import com.aws.greengrass.util.FileSystemPermission.Option;
 import com.aws.greengrass.util.Utils;
@@ -94,7 +94,7 @@ public abstract class Platform implements UserPlatform {
 
     public abstract SystemResourceController getSystemResourceController();
 
-    public abstract Exec createNewProcessRunner();
+    public abstract ExecBase createNewProcessRunner();
 
     /**
      * Set permissions on a path.
